fix bugs found during E2E testing, rename files and resolve minor conflicts
