+++ frontend/rebu/src/components/Home.jsx
@@ -1,55 +1,47 @@
-import {useState, Fragment} from 'react';
-import Map from './Map';
-import UserSummary from './UserSummary';
+import React, { useEffect, useState } from "react";
 
-import Chat from "./Chat/Chat";
+import HomeRider from "./HomeRider";
+import HomeDriver from "./HomeDriver";
+import ChatModal from './ChatModal';
+import HomeNotLoggedIn from './HomeNotLoggedIn';
 
-export default function Home(props) {
-  // changes to true, when the driver accepts the request
-  const [travelTD, settravelTD] = useState({
-    time: "0",
-    distance: "0",
-  })
-  // Coords
-  const[origin, setOrigin] = useState({lat:43.6453, lng:-79.3806});
-  const[destination, setDestination] = useState({lat:43.6706, lng:-79.3865});
+import Axios from "axios";
 
-  // Formated Addresses
-  const [startAddress, setstartAddress] = useState("");
-  const [finishAddress, setfinishAddress] = useState("");
+import './Home.scss';
 
-  
+const Homee = () => {
+  const [ user, setUser ] = useState({});
+  const token = localStorage.getItem("token");
 
-  console.log("USER INFO:", props.user)
-  console.log("USER NAME:", props.user.full_name)
+  const [ chatSelected, setChatSelected ] = useState(false);
+
+  useEffect(() => {
+    Axios.get("http://localhost:3001/users/data", { headers: { "x-access-token": token} })
+      .then((res) => {
+        setUser(res.data);
+      })
+      .catch(err => console.log(err));
+  }, []);
+
   return (
-    <Fragment>
-      <div className = "map">
-        <Map 
-          travelTD = {travelTD} settravelTD = {settravelTD} 
-          origin = {origin} setOrigin = {setOrigin} 
-          destination = {destination} setDestination = {setDestination}
-          startAddress = {startAddress} setstartAddress = {setstartAddress}
-          finishAddress = {finishAddress} setfinishAddress = {setfinishAddress}
+    <div id="homepage">
+      {!token && <HomeNotLoggedIn />}
+      {!user.driver && token &&
+        <HomeRider 
+          user={user} 
+          chatSelected={chatSelected}
         />
-      </div>
-      <div id="chat">
-        { props.chatSelected && 
-          <Chat 
-            name={props.user.full_name}
-            driver={props.user.driver}
-            user={props.user}
-          />
-        }
-      </div>
-        <UserSummary 
-          travelTD = {travelTD} settravelTD = {settravelTD}
-          origin = {origin} setOrigin = {setOrigin} 
-          destination = {destination} setDestination = {setDestination}
-          startAddress = {startAddress} 
-          finishAddress = {finishAddress} setfinishAddress = {setfinishAddress}
+      }
+      {user.driver && token &&
+        <HomeDriver 
+          user={user} 
+          setChatSelected={setChatSelected} 
+          chatSelected={chatSelected} 
         />
-    </Fragment>
-
+      }
+      <ChatModal setChatSelected={setChatSelected} chatSelected={chatSelected} />
+    </div>
   )
-}
+};
+
+export default Homee;
